+++ src/lib/radar/remap_ftn/timconv.f
@@ -43,7 +43,7 @@ cdis
 !       First generate a9time
         character*9 a9
 
-        integer*4 imon_a(12)
+        integer imon_a(12)
 
         data imon_a/0,31,59,90,120,151,181,212,243,273,304,334/
